Eliminate size_t to int potential loss warning by using size_t type instead of int.

+++ OpenSim/Auxiliary/auxiliaryTestFunctions.h
@@ -88,7 +88,7 @@ void CHECK_STORAGE_AGAINST_STANDARD(const OpenSim::Storage& result,
     std::vector<double> comparisons;
     result.compareWithStandard(standard, columnsUsed, comparisons);
 
-    int ncolumns = columnsUsed.size();
+    size_t ncolumns = columnsUsed.size();
 
     ASSERT(ncolumns > 0, testFile, testFileLine, 
            errorMessage + "- no common columns to compare!");
